Errors

* feat(nfts): endpoints now working with error checking

* feat: add multiple error mappers for exceptions

+++ service/src/main/java/ar/edu/itba/paw/service/FavoriteServiceImpl.java
@@ -35,8 +35,10 @@ public class FavoriteServiceImpl implements FavoriteService{
     @Transactional
     @Override
     public void removeNftFavorite(int productId, User user) {
-        Nft nft = nftService.getNFTById(productId).orElseThrow(NftNotFoundException::new);
-        favoriteDao.removeNftFavorite(nft, user);
+        if(isNftFavedByUser(user.getId(), productId)) {
+            Nft nft = nftService.getNFTById(productId).orElseThrow(NftNotFoundException::new);
+            favoriteDao.removeNftFavorite(nft, user);
+        }
     }
 
     @Override
